[SafeBox] When we get the financial information of a payment method of a safe box, use double instead of long numbers
This avoid problem in the rounding.

+++ src/org/openbravo/retail/posterminal/SafeBoxes.java
@@ -138,11 +138,11 @@ public class SafeBoxes extends JSONProcessSimple {
 
           // Save values in result JSON
           safeBoxPaymentMethodCounting.put("depositBalance",
-              safeBoxPaymentMethodTransaction.getLong("depositBalance"));
+              safeBoxPaymentMethodTransaction.getDouble("depositBalance"));
           safeBoxPaymentMethodCounting.put("paymentBalance",
-              safeBoxPaymentMethodTransaction.getLong("paymentBalance"));
+              safeBoxPaymentMethodTransaction.getDouble("paymentBalance"));
           safeBoxPaymentMethodCounting.put("initialBalance",
-              safeBoxPaymentMethodAccount.getLong("initialBalance"));
+              safeBoxPaymentMethodAccount.getDouble("initialBalance"));
 
           safeBoxPaymentMethod.put("safeBoxCounting", safeBoxPaymentMethodCounting);
         }
